linker : apply relocations

+++ linker/reloc.c
@@ -11,7 +11,7 @@
 
 int reloc(struct elf_object *object, Elf_Rela *rel) {
 	uintptr_t sym_val = 0;
-	int sym_index = ELF_R_SYM(rel->r_info);
+	size_t sym_index = ELF_R_SYM(rel->r_info);
 	if (sym_index != STN_UNDEF) {
 		if (sym_index >= object->symbols_count) {
 			dl_error("invalid symbol index");
